common/playlist: make playlist_entry_from_index argument int64_t
This prevents overflow when specifying very large index values
in later commits.

+++ common/playlist.c
@@ -380,7 +380,7 @@ int playlist_entry_count(struct playlist *pl)
 
 // Return entry for which playlist_entry_to_index() would return index.
 // Return NULL if not found.
-struct playlist_entry *playlist_entry_from_index(struct playlist *pl, int index)
+struct playlist_entry *playlist_entry_from_index(struct playlist *pl, int64_t index)
 {
     return index >= 0 && index < pl->num_entries ? pl->entries[index] : NULL;
 }
